refactor(server): move route request helpers into package
